Restructure tests
